refactor Matrix class def
