--- test/TestMatrix/TestMatrixDecomposition.cpp
@@ -1,229 +1,0 @@
-﻿// L=============================================================================
-// L This software is distributed under the MIT license.
-// L Copyright 2021 Péter Kardos
-// L=============================================================================
-
-#pragma warning(disable : 4244)
-
-#include "../TestGenerators.hpp"
-
-#include <Mathter/Common/Approx.hpp>
-#include <Mathter/Matrix.hpp>
-
-#include <catch2/catch_approx.hpp>
-#include <catch2/catch_template_test_macros.hpp>
-#include <catch2/catch_test_macros.hpp>
-#include <complex>
-
-
-
-using namespace mathter;
-using Catch::Approx;
-
-
-using TypeListFloating = TestTypeList<TypesFloating, PackedAll, OrdersAll, LayoutsAll>;
-
-
-TEMPLATE_LIST_TEST_CASE("Matrix - LU decomposition", "[Matrix]", TypeListFloating) {
-	SECTION(TestType::Name()) {
-		using M = typename TestType::template Matrix<3, 3>;
-		M A = {
-			3, -0.1, -0.2,
-			0.1, 7, -0.3,
-			0.3, -0.2, 10
-		};
-
-		auto [L, U] = DecomposeLU(A);
-
-		for (int i = 0; i < A.RowCount(); ++i) {
-			for (int j = 0; j < i - 1; ++j) {
-				REQUIRE(U(i, j) == Approx(0.0));
-				REQUIRE(L(j, i) == Approx(0.0));
-			}
-		}
-
-		auto Mprod = L * U;
-		REQUIRE(ApproxVec(A) == Mprod);
-	}
-}
-
-
-TEMPLATE_LIST_TEST_CASE("Matrix - LU solve", "[Matrix]", TypeListFloating) {
-	SECTION(TestType::Name()) {
-		using M = typename TestType::template Matrix<3, 3>;
-		using V = typename TestType::template Vector<3>;
-		M A = {
-			3, -0.1f, -0.2f,
-			0.1f, 7, -0.3f,
-			0.3f, -0.2f, 10
-		};
-		V b = { 7.85, -19.3, 71.4 };
-		V x;
-		V xexp = { 3, -2.5, 7 };
-
-		x = DecomposeLU(A).Solve(b);
-		REQUIRE(ApproxVec(x) == xexp);
-	}
-}
-
-
-TEMPLATE_LIST_TEST_CASE("Matrix - LUP decomposition", "[Matrix]", TypeListFloating) {
-	SECTION(TestType::Name()) {
-		using M = typename TestType::template Matrix<3, 3>;
-
-		M A = {
-			3, -0.1f, -0.2f,
-			0.3f, -0.2f, 10,
-			0.1f, 7, -0.3f
-		};
-
-		auto [L, U, P] = DecomposeLUP(A);
-
-		for (int i = 0; i < A.RowCount(); ++i) {
-			for (int j = 0; j < i - 1; ++j) {
-				REQUIRE(U(i, j) == Approx(0.0f));
-				REQUIRE(L(j, i) == Approx(0.0f));
-			}
-		}
-
-		M Pm = Zero();
-		for (int i : P) {
-			Pm(i, P(i)) = 1.0f;
-		}
-
-		auto Mprod = Transpose(Pm) * L * U;
-		REQUIRE(ApproxVec(A) == Mprod);
-	}
-}
-
-
-TEMPLATE_LIST_TEST_CASE("Matrix - LUP solve", "[Matrix]", TypeListFloating) {
-	SECTION(TestType::Name()) {
-		using M = typename TestType::template Matrix<4, 4>;
-		using V = typename TestType::template Vector<4>;
-
-		M A = {
-			1, 3, 4, 6,
-			3, 6, 2, 6,
-			9, 2, 6, 7,
-			6, 2, 7, 5
-		};
-		V b = { 3, 4, 2, 8 };
-		V x;
-		V xexp = { -94.f / 497, 895.f / 497, 1000.f / 497, -850.f / 497 };
-
-		x = DecomposeLUP(A).Solve(b);
-
-		REQUIRE(ApproxVec(x) == xexp);
-	}
-}
-
-
-TEST_CASE("Matrix - LUP decomposition singular", "[Matrix]") {
-	Matrix<float, 3, 3> A = {
-		1, 0, 0,
-		0, 0, 1,
-		0, -1, 0
-	};
-
-	auto [L, U, P] = DecomposeLUP(A);
-
-	for (int i = 0; i < A.RowCount(); ++i) {
-		for (int j = 0; j < i - 1; ++j) {
-			REQUIRE(U(i, j) == Approx(0.0f));
-			REQUIRE(L(j, i) == Approx(0.0f));
-		}
-	}
-
-	Matrix<float, 3, 3> Pm = Zero();
-	for (int i : P) {
-		Pm(i, P(i)) = 1.0f;
-	}
-
-	auto Mprod = Transpose(Pm) * L * U;
-	REQUIRE(ApproxVec(A) == Mprod);
-}
-
-
-TEMPLATE_LIST_TEST_CASE("Matrix - QR decomposition", "[Matrix]", TypeListFloating) {
-	SECTION(TestType::Name()) {
-		using M33 = typename TestType::template Matrix<3, 3>;
-		using M54 = typename TestType::template Matrix<5, 4>;
-		using M45 = typename TestType::template Matrix<4, 5>;
-
-		// example from wikipedia SVD article
-		M54 A1 = Transpose(M45{
-			1, 0, 0, 1,
-			2, 0, 0, 3,
-			0, 0, 0, 0,
-			0, 0, 0, 0,
-			2, 0, 0, 0 });
-		auto [Q1, R1] = DecomposeQR(A1);
-		M54 A1assembled = Q1 * R1;
-		REQUIRE(ApproxVec(A1assembled) == A1);
-
-
-		// the same matrix as the LU
-		M33 A2 = {
-			3, -0.1f, -0.2f,
-			0.1f, 7, -0.3f,
-			0.3f, -0.2f, 10
-		};
-
-		auto [Q2, R2] = DecomposeQR(A2);
-
-		M33 A2assembled = Q2 * R2;
-		REQUIRE(ApproxVec(A2assembled) == A2);
-	}
-}
-
-
-TEMPLATE_LIST_TEST_CASE("Matrix - SVD", "[Matrix]", TypeListFloating) {
-	SECTION(TestType::Name()) {
-		using M33 = typename TestType::template Matrix<3, 3>;
-		using M54 = typename TestType::template Matrix<5, 4>;
-		using M45 = typename TestType::template Matrix<4, 5>;
-
-		// example from wikipedia SVD article
-		M54 A1 = Transpose(M45{
-			1, 0, 0, 1, 2,
-			0, 0, 3, 0, 0,
-			0, 0, 0, 0, 0,
-			0, 2, 0, 0, 0 });
-
-		auto [U1, S1, V1] = DecomposeSVD(A1);
-		auto A1assembled = U1 * S1 * V1;
-		REQUIRE(ApproxVec(A1) == A1assembled);
-
-		auto [U1T, S1T, V1T] = DecomposeSVD(Transpose(A1));
-		auto A1Tassembled = U1T * S1T * V1T;
-		REQUIRE(ApproxVec(A1Tassembled) == Transpose(A1));
-
-		// The same matrix as the LU
-		M33 A2 = {
-			3, -0.1f, -0.2f,
-			0.1f, 7, -0.3f,
-			0.3f, -0.2f, 10
-		};
-
-		auto [U2, S2, V2] = DecomposeSVD(A2);
-		auto A2assembled = U2 * S2 * V2;
-		REQUIRE(ApproxVec(A2assembled) == A2);
-	}
-}
-
-
-TEST_CASE("Matrix - SVD Identity", "[Matrix]") {
-	Matrix<float, 2, 2> m = Identity();
-
-	auto svd = DecomposeSVD(m);
-	REQUIRE(svd.U == ApproxVec(Matrix<float, 2, 2>(Identity())));
-	REQUIRE(svd.S == ApproxVec(Matrix<float, 2, 2>(Identity())));
-	REQUIRE(svd.V == ApproxVec(Matrix<float, 2, 2>(Identity())));
-
-	Matrix<float, 4, 4> m4 = Identity();
-	auto svd4 = DecomposeSVD(m4);
-	REQUIRE(svd4.U == ApproxVec(Matrix<float, 4, 4>(Identity())));
-	REQUIRE(svd4.S == ApproxVec(Matrix<float, 4, 4>(Identity())));
-	REQUIRE(svd4.V == ApproxVec(Matrix<float, 4, 4>(Identity())));
-}
